Use a constant-time load balancer (#266)
tower-rs/tower#288 and tower-rs/tower#293 changed the Balance
implementation substantially. Now, a new layer, `spawn_ready` is
inserted around endpoint stack to ensure that readiness is driven
on a background task.

In support of this change, the `pending` layer was removed from the
endpoint stack and, instead, the discovery system is now responsible for
driving pending services to be materialized.