Properly fall back in the dst_router (#291)
### Motivation

TLDR: Requests whose target authority is a socket address do not properly fall
back. Requests to unresolvable authorities which are DNS names do fall back as
expected.

In a control plane branch, I have the tap controller properly adding the header
metadata to tap requests, and the pod proxy’s expecting the tap identity the
requests would be coming from. After installing a local build of this control
plane and the proxy that supports `l5d-require-id`, I noticed that the tap
controller’s proxy was still not setting identity on `outbound::Endpoint`s when
the header was there.

I began looking for the `debug!` lines that would be hit in the
`orig_dst_router` because I would expect to be hitting that router, and did not
see them—indicating those requests were still going through the balancer.

It turns out that requests with an authority that we cannot resolve still get a
single slot in the balancer, and therefore do not fallback because they do not
return the `Unresolvable` error. We match on `Socket` [here](https://github.com/linkerd/linkerd2-proxy/blob/master/src/app/outbound.rs#L207) and make a new
slot in the balancer [here](https://github.com/linkerd/linkerd2-proxy/blob/master/src/app/outbound.rs#L276)

This is an issue because if these requests contain the `l5d-require-id` header,
the `Endpoint` `identity` is not set because we do not go through the
`orig_dst_router` and therefore do not have a TLS connection

### Solution

In `outbound::discover`, `Resolve` will now only attempt to resolve
`Addr::Name`, and return a `Unresolvable` error on `Addr::Socket`. For requests
that are `Addr::Socket`, we rely on getting the original dst of the request and
therefore routes through `orig_dst_router`.

### Tests

The tests that specifically test SO_ORIGINAL_DST have been removed because those
tests now result in errors which is the expected behavior now.

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>