metrics: add `request_handle_us` histogram (#294)
As initially described in linkerd/linkerd2#730, there's a general need
to understand the proxy's latency overhead. We can begin by recording
the time a request spends in the proxy. 

linkerd/linkerd2#3098 proposes adding a new `request_handle_us`
histogram to the proxy with only a `direction` label. This histogram 
should store the elapsed time (in microseconds) from the moment a 
request reaches the source stack until the request is dropped.

This branch adds the `request_handle_us` histogram to the proxy's 
metrics. Handle time is recorded using a `Tracker` type which is 
inserted into each request's extensions map by a new layer at the
top of the stack. Each tracker has a reference count, and when 
requests are cloned for retries, the tracker is cloned into the new
request as well, incrementing the reference count. The reference 
count is decremented when trackers are dropped, and when it reaches
zero, the elapsed time is recorded.

Since this depends on a time measurement that's internal to the proxy,
it's difficult to write reliable tests for it. Unlike the latency
histograms, we cannot use a test service that waits for a period of time
to ensure that the recorded latencies have a lower bound. Instead, I did
some manual verification that the new metric is present in the proxy's
metrics endpoint, that its total count is equal to the number of
requests I've sent through the proxy, and that the handle time
measurements appear to be less than the total request latency
measurements.

Closes linkerd/linkerd2#3098 

Signed-off-by: Eliza Weisman <eliza@buoyant.io>