tests: Properly simulate destination errors (#332)
The `discovery` tests attempt to simulate the scenario where the destination
service responds with an `InvalidArgument` gRPC status so the proxy
knows to skip balancing. The tests don't accurately simulate this
behavior, however.

The [destination controller eagerly fails such requests][1], rather than
successfully returning a stream that fails immediately. At the protocol
level, this causes the `grpc-status` to be encoded in the firt and only
HTTP response header frame, rather than in a trailer (i.e after the
initial `END_HEADERS` frame was sent indicating the begining of the
respnose steram); and in tower, this causes the initial response future
to fail.

This change modifies the mock controller and discovery tests so that
tests can fail the response eagerly without initiating the response
stream.

[1]: https://github.com/linkerd/linkerd2/blob/e281fb3410138a8a9505220ecddea627dd9fdc2d/controller/api/destination/server.go#L96-L112