core: Introduce the Recover trait (#347)
* core: Introduce the Recover trait

The `Recover` trait provides a means to inspect an error and, if it is
non-fatal, apply a backoff strategy. Backoff strategies are implemented
as a `Stream` that indicates when an operation may be
resumed/reattempted.

This affords an opportunity to make the `proxy::reconnect` module
decoupled from its backoff strategy (which will more easily allow the
strategy to be reused in other contexts, i.e. to support #341).

This change also corrects a logic error in the exponential backoff
strategy that prevented jitter from being applied. The code has been
refactored slightly to improve testability.

There was some unknown interaction (possibly a bug in tower-reconect?)
that causes test failures related to reconnect--where the task is
sometimes not notified as it should be.

This change replaces use of tower-reconnect with a `Recover`-aware
state machine and eliminates that dependency.

Finally, the `linkerd2-error` crate extracts the `Recover, `Error`, and
`Never` types. And, as this allows some crates to stop depending on
proxy-core, they have been renamed to not include a "proxy" prefix.