Split service discovery into composable components (#341)
The `linkerd2_proxy_resolve` crate, which was split out of the main
application in #318, contains a monolithic `Resolve` implementation.
Over time, this code has evolved in ways that have made it difficult to
modify and extend, especially and functionality cannot be tested
independently.

This change reorganizes the proxy's service discovery logic into a few
composable components. The `Resolve` API has been updated to facilitate
this.

Service discovery consists of two primary traits: `Resolve` and
`Discover`. `Resolve` is used within the proxy to model discovery
sources like the Destination API and DNS; and `Discover` updates a load
balancer with changes to its replica pool.

The former `Resolve` trait has been replaced with a trait alias for
`Service`, which allows for back-pressure, etc. Furthermore, the API has
been updated so that it is easier to gracefully recover after a
resolution stream is lost (by changing `Update` to contain a `Vec`).

Now that `Resolve` is implemented as a `Service`, it becomes
straightforward to model service discovery in terms of a stack of
services, as we do throughout the project:

* The `linkerd2-proxy-api-resolve` crate provides an implementation of
  `Resolve` backed by the destination API. It does NOT handle
  reconnects, buffering/spawning, etc.
* The `linkerd2-request-filter` crate provides a simple
  middleware that can fail requests based on the request. This is used to
  implement the destination client's name/suffix restrictions.
* The `linkerd2-proxy-resolve` crate provides `Resolve`-middlewares that
  implement error recovery/backoff and endpoint-type-mapping.
* The `linkerd2-proxy-discover crate provides a suite of middlewares
  that facilitate powering a balancer with a `Resolve`. Specifically:
  * `buffer` ensures that discovery updates are processed even when the
    balancer is not polling for updates.
  * `from_resolve` wraps a `Resolve` to produce `Discover` response.
  * `make_endpoint` wraps a `Discover`-producing service to build each
    inserted endpoint into a Service.

Almost all of this logic was happening in a single module, coupled to
our gRPC interface. Now, these components may be easily re-purposed (for
instance, in order to back discovery by DNS).