Make fallback HTTP-agnostic (#393)
The fallback module is currently HTTP-aware, as it must make response
body types uniform.

This change modifies fallback to instead require that both _service
types_ are identical. This is accomplished by boxing the services, using
a new http boxed service type.

This has the side-effect of dramatically improving compilation times for
the outbound proxy.