stack: Introduce the Proxy trait (#441)
The `stack::Proxy` trait models middlewares that may modify the request
or response types but do not need to participate in the inner Service's
readiness.

This will be an important component of a new `service-profiles`
implementation: each request route does not need an independent clone of
the underlying concrete service (i.e., that it must drive to readiness
independently). Instead, the profile router can be responsible for
driving a single concrete router to readiness and invoke the ready
service through the appropriate route's `Proxy`.

Implementations will follow.