Introduce a backpressure-propagating buffer (#451)
The tower buffer implementation uses the underlying channel for
backpressure. This admits requests before the inner service becomes
ready, and doesn't fully propagate error conditions.

This replacement buffer implementation exposes readiness via a `Watch`.
This prevents callers from sending requests when the inner service is
not ready and allows service errors to be propagated reliably.  The
buffer's background task drops the inner service as soon as it fails.

The effect of this is that the buffer's channel is only used for
requests that race to submit requests while the inner service is ready,
and so it seems likely that they only need enough slots to satisfy the
proxy's expected concurrency (i.e. in terms of cores).