Add a `mock-orig-dst` feature flag (#466)
The proxy code is generic over the strategy it uses for determining
the original destination address for a given connection. In integration
tests, the strategy is replaced with a configured version.

The `mock-orig-dst` feature modifies the implementation to be
swappable at compile-time via feature flag. This removes integration-
test-specific test code, and makes it generally possible to start the
proxy outside of Kubernetes (via the compile-time flag). When this
feature is enabled, runtime configuration is required via a pair of new
environment variables.

There was an integration test that tested the proxy's behavior when
an address could not be determined, but that test has been removed
as that use case is not particularly important.

To build a proxy in this mode, one runs:
```
:; (cd linkerd2-proxy && cargo build --release --features=mock-orig-dst)
```