integration: fix missing traces (#572)
Currently, there are a couple of issues that prevent traces emitted by
the proxy's integration tests from being reported correctly. Many of our
tests begin with:

```rust
let _ = trace_init();
```

This doesn't _actually_ do what we want, for two reasons:

* `trace_init` doesn't actually set a trace subscriber as the default,
  it just _constructs_ one,
* `let _` drops the value _immediately_.

So, this code is constructing a subscriber and then immediately dropping
it, which does nothing.

In practice, we mostly don't notice this, since the test support
servers, clients, and proxy run in separate threads with their _own_
trace subscribers, which are set up correctly, and very few of the tests
themselves actually emit traces. So even though this does the wrong
thing, everything basically works. However, if the dedicated test
support threads are removed and the test support stuff moved to spawned
tasks in the test functions, all the traces go away.

This PR fixes that by making `trace_init` actually set a default
subscriber, adding a new `trace_subscriber` function that just
constructs the subscriber and level handle (which is necessary to run a
test proxy), and changing the tests so they don't drop their guards
until the test ends.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>
Co-authored-by: Kevin Leimkuhler <kevin@kleimkuhler.com>
Co-authored-by: Oliver Gould <ver@buoyant.io>