Make FailFast cloneable (#617)
Currently, the FailFast middleware is not cloneable, as it maintains
state including a timeout future.

This change modifies FailFast to implement Clone. When the module is
cloned, it reverts to the open state.

This has no impact on the current proxy, but when the stack gets cloned
(as required by an upcoming change), this will have the effect that
connections do not share failfast state. This seems better than adding a
Lock or Buffer just to accomodate this.