Decouple discovery buffering from endpoint conversion (#631)
The discovery layer handles both building a service for each discovered
endpoint, as well as a buffer that fails when the service is not polled
(i.e. because it has been leaked somehow).

This coupling doesn't match other use cases though when, for instance,
the client may not be used frequently, as is the case for
proxy-to-control-plane communication.

This change decouples these layers so that the buffering is decoupled
from endpoint-service construction.