outbound: Clarify Http target types (#653)
The outbound `endpoint::Target` type is generic to the point of being
useless.

This change modifies the stack target types to be either `HttpLogical`,
`HttpConcrete`, or `HttpEndpoint`. Furthermore, the `canonicalize`
middleware no longer uses a bespoke trait, instead opting for
`AsRef` + `AsMut`.

This sets up further changes to the outbound HTTP stack.