Only allow name-based profile discovery for inbound requests (#695)
The proxy uses a single profile client for both inbound and outbound
resolutions. We now want to perform profile discovery for all outbound
traffic; but this is overkill for inbound traffic. We should only do
profile lookup for traffic that includes a valid service name and not,
for instance, healthchecks and prometheus scrapes.

This changes inbound to only do resolution for valid names, skipping
traffic that is not targeting a valid service name.

In order to do this, the following has changed:

- An `AddrMatch` utility is introduced to `core` so that this logic need
  not be propagated.
- `app::dst` now simply builds a profile client without filtering or
  recovery.
- The profile client is no longer responsible for synthesizing a
  default profile on timeout -- this will be extracted into the stack in
  a followup.
- The `profiles::discovery::layer` utility now applies filtering and
  discovery, so that the inbound and outbound proxies may supply their own
  filters.
- When we reject discovery and handle failed discovery, we always use gRPC
 error codes now