tracing: Fix time formatting to ensure alignment (#722)
In logs, we emit timestamps as time fractional seconds since the process
started. These timestamps are intended to be formatted at a fixed
length, but we don't properly truncate nanoseconds to fit the 6-digit
field we leave for subsecond time.

This change converts nanoseconds to microseconds when formatting
the uptime so that the value fits into six digits. It also adds a test
to validate this behavior.