tracing: Support disabling tracing entirely (#723)
When the proxy log level is set to `off`, the proxy now skips all
tracing initialization and disables the tracing admin endpoints. This
allows us to minimize the overhead of tracing in benchmarks and
performance-critical environments.

This has a noticeable impact on _memory_. In local tests, we see
`main` use around ~24MB with the log level set to `off`. With this
branch, memory comes down to ~16MB.