h2: enable HTTP/2 keepalive PING frames (#737)
This branch enables HTTP/2 PING frames in the proxy's HTTP/2 clients and
servers. The timeout for responding to a PING frame is configured based
on the `{INBOUND, OUTBOUND}_{CONNECT, ACCEPT}_KEEPALIVE` env variables,
and the interval between PING frames is currently 1/4th of the timeout.
I'm happy to change that if anyone has better ideas.

Collecting metrics related to H2 PINGs probably requires support in 
Hyper that doesn't currently exist, so this PR doesn't add that. We 
can implement metrics in a follow-up, as it's lower priority.

Closes linkerd/linkerd2#1580